pttc: fix 'implicit conversion changes signedness' warning
Clang complains about an implicit conversion from int to size_t to
change the signedness.  Fix it.

+++ pttc/src/parse.c
@@ -279,7 +279,7 @@ static char *expfilename(struct parser *p, const char *extra)
 
 	pfname[len] = 0;
 
-	return memcpy(pfname, filename, len);
+	return memcpy(pfname, filename, (size_t) len);
 }
 
 /* Returns true if @c is part of a label; false otherwise.  */
